migrate commands into io/commands and keep public facing command file
