+++ src/subhkl/io/commands/finder.py
@@ -1,0 +1,146 @@
+import h5py
+
+from subhkl.integration import Peaks
+
+def run(
+    filename: str,
+    instrument: str,
+    output_filename: str = "output.h5",
+    finder_algorithm: str = "peak_local_max",
+    show_progress: bool = True,
+    create_visualizations: bool = False,
+    show_steps: bool = False,
+    peak_local_max_min_pixel_distance: int = -1,
+    peak_local_max_min_relative_intensity: float = -1,
+    peak_local_max_normalization: bool = False,
+    mask_file: str | None = None,
+    mask_rel_erosion_radius: float | None = None,
+    thresholding_noise_cutoff_quantile: float = 0.8,
+    thresholding_min_peak_dist_pixels: float = 8.0,
+    thresholding_blur_kernel_sigma: int = 5,
+    thresholding_open_kernel_size_pixels: int = 3,
+    wavelength_min: float | None = None,
+    wavelength_max: float | None = None,
+    region_growth_distance_threshold: float = 1.5,
+    region_growth_minimum_sigma: float | None = None,
+    region_growth_minimum_intensity: float = 4500.0,
+    region_growth_maximum_pixel_radius: float = 17.0,
+    peak_center_box_size: int = 15,
+    peak_smoothing_window_size: int = 15,
+    peak_minimum_pixels: int = 30,
+    peak_minimum_signal_to_noise: float = 1.0,
+    peak_pixel_outlier_threshold: float = 2.0,
+    sparse_rbf_alpha: float = 0.1,
+    sparse_rbf_gamma: float = 1.0,
+    sparse_rbf_min_sigma: float = 1.5,
+    sparse_rbf_max_sigma: float = 10.0,
+    sparse_rbf_max_peaks: int = 500,
+    sparse_rbf_chunk_size: int = 512,
+    sparse_rbf_tile_rows: int = 2,
+    sparse_rbf_tile_cols: int = 2,
+    sparse_rbf_loss: str = "gaussian",
+    sparse_rbf_auto_tune_alpha: bool = False,
+    sparse_rbf_candidate_alphas: str = "3.0,5.0,10.0,15.0,20.0,25.0,30",
+    max_workers: int = 16,
+):
+    print(f"Creating peaks from {filename} for instrument {instrument}")
+
+    wavelength_kwargs = {}
+    if wavelength_min:
+        wavelength_kwargs["wavelength_min"] = wavelength_min
+    if wavelength_max:
+        wavelength_kwargs["wavelength_max"] = wavelength_max
+
+    peaks = Peaks(filename, instrument, **wavelength_kwargs)
+
+    peak_kwargs = {"algorithm": finder_algorithm}
+    if finder_algorithm == "peak_local_max":
+        if peak_local_max_min_pixel_distance > 0:
+            peak_kwargs["min_pix"] = peak_local_max_min_pixel_distance
+        if peak_local_max_min_relative_intensity > 0:
+            peak_kwargs["min_rel_intensity"] = peak_local_max_min_relative_intensity
+        peak_kwargs["normalize"] = peak_local_max_normalization
+    elif finder_algorithm == "thresholding":
+        peak_kwargs.update(
+            {
+                "noise_cutoff_quantile": thresholding_noise_cutoff_quantile,
+                "min_peak_dist_pixels": thresholding_min_peak_dist_pixels,
+                "blur_kernel_sigma": thresholding_blur_kernel_sigma,
+                "open_kernel_size_pixels": thresholding_open_kernel_size_pixels,
+                "show_steps": show_steps,
+                "show_scale": "log",
+            }
+        )
+    elif finder_algorithm == "sparse_rbf":
+        # Because we separated Typer from core logic, this split is 100% safe
+        alpha_list = [float(k.strip()) for k in sparse_rbf_candidate_alphas.split(",")]
+
+        peak_kwargs.update(
+            {
+                "alpha": sparse_rbf_alpha,
+                "gamma": sparse_rbf_gamma,
+                "min_sigma": sparse_rbf_min_sigma,
+                "max_sigma": sparse_rbf_max_sigma,
+                "max_peaks": sparse_rbf_max_peaks,
+                "chunk_size": sparse_rbf_chunk_size,
+                "show_steps": show_steps,
+                "show_scale": "linear",
+                "tiles": (sparse_rbf_tile_rows, sparse_rbf_tile_cols),
+                "loss": sparse_rbf_loss,
+                "auto_tune_alpha": sparse_rbf_auto_tune_alpha,
+                "candidate_alphas": alpha_list,
+            }
+        )
+    else:
+        raise ValueError("Invalid finder algorithm")
+
+    peak_kwargs.update(
+        {
+            "mask_file": mask_file,
+            "mask_rel_erosion_radius": mask_rel_erosion_radius,
+        }
+    )
+
+    integration_params = {
+        "region_growth_distance_threshold": region_growth_distance_threshold,
+        "region_growth_minimum_sigma": region_growth_minimum_sigma,
+        "region_growth_minimum_intensity": region_growth_minimum_intensity,
+        "region_growth_maximum_pixel_radius": region_growth_maximum_pixel_radius,
+        "peak_center_box_size": peak_center_box_size,
+        "peak_smoothing_window_size": peak_smoothing_window_size,
+        "peak_minimum_pixels": peak_minimum_pixels,
+        "peak_minimum_signal_to_noise": peak_minimum_signal_to_noise,
+        "peak_pixel_outlier_threshold": peak_pixel_outlier_threshold,
+    }
+
+    detector_peaks = peaks.get_detector_peaks(
+        peak_kwargs,
+        integration_params,
+        visualize=create_visualizations,
+        show_progress=show_progress,
+        file_prefix=filename,
+        max_workers=max_workers,
+    )
+
+    peaks.write_hdf5(
+        output_filename=output_filename,
+        detector_peaks=detector_peaks,
+        instrument_wavelength=[peaks.wavelength.min, peaks.wavelength.max],
+    )
+
+    # copy over cell params
+    copy_keys = [
+        "sample/a",
+        "sample/b",
+        "sample/c",
+        "sample/alpha",
+        "sample/beta",
+        "sample/gamma",
+        "sample/space_group",
+    ]
+
+    with h5py.File(output_filename, "a") as f:
+        with h5py.File(filename, "r") as f_in:
+            for key in copy_keys:
+                if key in f_in:
+                    f_in.copy(f_in[key], f, key)
