+++ src/subhkl/io/commands/index.py
@@ -1,0 +1,582 @@
+import h5py
+import numpy as np
+
+from subhkl.instrument.goniometer import (
+    get_rotation_data_from_nexus,
+)
+
+from subhkl.integration import Peaks
+from subhkl.optimization import FindUB
+
+from subhkl.instrument.detector import calibrate_from_file
+def run(
+    peaks_h5_filename: str,
+    output_peaks_filename: str,
+    a: float | None = None,
+    b: float | None = None,
+    c: float | None = None,
+    alpha: float | None = None,
+    beta: float | None = None,
+    gamma: float | None = None,
+    space_group: str | None = None,
+    wavelength_min: float | None = None,
+    wavelength_max: float | None = None,
+    ki_vec: list[float] | np.ndarray | None = None,
+    original_nexus_filename: str | None = None,
+    instrument_name: str | None = None,
+    strategy_name: str = "DE",
+    sigma_init: float | None = None,
+    n_runs: int = 1,
+    population_size: int = 1000,
+    gens: int = 100,
+    seed: int = 0,
+    tolerance_deg: float = 0.1,
+    freeze_orientation: bool = False,
+    refine_lattice: bool = False,
+    lattice_bound_frac: float = 0.05,
+    refine_goniometer: bool = False,
+    refine_goniometer_axes: list[str] | None = None,
+    goniometer_bound_deg: float = 5.0,
+    refine_sample: bool = False,
+    sample_bound_meters: float = 0.005,
+    refine_beam: bool = False,
+    beam_bound_deg: float = 1.0,
+    refine_detector: bool = False,
+    refine_detector_banks: list[int] | None = None,
+    detector_modes: list[str] | None = None,
+    detector_trans_bound_meters: float = 0.005,
+    detector_rot_bound_deg: float = 1.0,
+    detector_global_rot_bound_deg: float = 2.0,
+    detector_global_rot_axis: list[float] | np.ndarray | None = None,
+    detector_global_trans_bound_meters: float = 0.01,
+    detector_radial_bound_frac: float = 0.05,
+    bootstrap_filename: str | None = None,
+    batch_size: int | None = None,
+    input_data: dict | None = None,
+    num_candidates: int | None = None,
+):
+    input_data = input_data or {}
+
+    if detector_modes is None:
+        detector_modes = ["independent"]
+    if detector_global_rot_axis is None:
+        detector_global_rot_axis = [0.0, 1.0, 0.0]
+
+    # --- INJECT BOOTSTRAP PHYSICS DIRECTLY ---
+    if bootstrap_filename:
+        calibrate_from_file(bootstrap_filename, instrument_name)
+        with h5py.File(bootstrap_filename, "r") as b_f:
+            if "sample/a" in b_f:
+                a = b_f["sample/a"][()]
+            if "sample/b" in b_f:
+                b = b_f["sample/b"][()]
+            if "sample/c" in b_f:
+                c = b_f["sample/c"][()]
+            if "sample/alpha" in b_f:
+                alpha = b_f["sample/alpha"][()]
+            if "sample/beta" in b_f:
+                beta = b_f["sample/beta"][()]
+            if "sample/gamma" in b_f:
+                gamma = b_f["sample/gamma"][()]
+
+    print(f"Loading peaks from: {peaks_h5_filename}")
+    with h5py.File(peaks_h5_filename, "r") as f:
+        if a is None:
+            a = f["sample/a"][()] if "sample/a" in f else None
+        if b is None:
+            b = f["sample/b"][()] if "sample/b" in f else None
+        if c is None:
+            c = f["sample/c"][()] if "sample/c" in f else None
+        if alpha is None:
+            alpha = f["sample/alpha"][()] if "sample/alpha" in f else None
+        if beta is None:
+            beta = f["sample/beta"][()] if "sample/beta" in f else None
+        if gamma is None:
+            gamma = f["sample/gamma"][()] if "sample/gamma" in f else None
+
+        if space_group is None:
+            file_sg = f["sample/space_group"][()] if "sample/space_group" in f else None
+            space_group = (
+                file_sg.decode("utf-8") if isinstance(file_sg, bytes) else file_sg
+            )
+
+        if None in (a, b, c, alpha, beta, gamma, space_group):
+            raise ValueError(
+                "Unit cell parameters (a,b,c,alpha,beta,gamma) and Space Group must be provided via CLI or exist in the input file."
+            )
+
+        from subhkl.core.spacegroup import get_space_group_object
+
+        try:
+            get_space_group_object(space_group)
+        except ValueError as e:
+            raise ValueError(f"Invalid space group '{space_group}': {e}")
+
+        if wavelength_min is None or wavelength_max is None:
+            if "instrument/wavelength" in f:
+                wl = f["instrument/wavelength"][()]
+                if wavelength_min is None:
+                    wavelength_min = float(wl[0])
+                if wavelength_max is None:
+                    wavelength_max = float(wl[1])
+            else:
+                raise ValueError(
+                    "Wavelength min/max not provided and not found in input file."
+                )
+
+        keys_to_load = [
+            "peaks/intensity",
+            "peaks/sigma",
+            "peaks/radius",
+            "goniometer/R",
+            "goniometer/axes",
+            "goniometer/angles",
+            "goniometer/names",
+            "files",
+            "file_offsets",
+            "peaks/run_index",
+            "peaks/image_index",
+            "bank",
+            "bank_ids",
+            "sample/offset",
+            "beam/ki_vec",
+            "peaks/pixel_r",
+            "peaks/pixel_c",
+        ]
+        for k in keys_to_load:
+            if k in f:
+                input_data[k] = f[k][()]
+
+        if ki_vec is not None:
+            ki_vec_val = np.array(ki_vec)
+        else:
+            ki_vec_val = (
+                f["beam/ki_vec"][()]
+                if "beam/ki_vec" in f
+                else np.array([0.0, 0.0, 1.0])
+            )
+
+        detector_params = None
+        peak_pixel_coords = None
+        target_banks = None
+
+        if "peaks/pixel_r" in f and "peaks/pixel_c" in f:
+            print("Reconstructing physical geometry from pixels for optimization...")
+            if not instrument_name or not original_nexus_filename:
+                raise ValueError(
+                    "ERROR: Finder file contains pixels. You must provide --instrument and --nexus to rebuild geometry."
+                )
+
+            pixel_r = f["peaks/pixel_r"][()]
+            pixel_c = f["peaks/pixel_c"][()]
+
+            bank_array = None
+            if "bank" in f:
+                bank_array = f["bank"][()]
+            elif "peaks/bank" in f:
+                bank_array = f["peaks/bank"][()]
+            elif "bank_ids" in f and "peaks/image_index" in f:
+                b_ids = f["bank_ids"][()]
+                img_idx = f["peaks/image_index"][()]
+                bank_array = np.array([b_ids[int(idx)] for idx in img_idx])
+            else:
+                bank_array = f["peaks/image_index"][()]
+
+            peaks_obj = Peaks(original_nexus_filename, instrument_name)
+            from subhkl.config import beamlines
+            from subhkl.instrument.detector import Detector
+
+            if refine_detector:
+                all_physical_banks = [int(k) for k in beamlines[instrument_name].keys()]
+                target_banks = (
+                    refine_detector_banks
+                    if refine_detector_banks
+                    else sorted(all_physical_banks)
+                )
+
+                centers, uhats, vhats, m, n, pw, ph = [], [], [], [], [], [], []
+                bank_to_idx = {}
+
+                for idx, b_id in enumerate(target_banks):
+                    try:
+                        det = peaks_obj.get_detector(b_id)
+                        centers.append(det.center)
+                        uhats.append(det.uhat)
+                        vhats.append(det.vhat)
+                        m.append(det.m)
+                        n.append(det.n)
+                        pw.append(det.width / det.m)
+                        ph.append(det.height / det.n)
+                        bank_to_idx[b_id] = idx
+                    except Exception as e:
+                        print(f"WARNING: Could not load geometry for bank {b_id}: {e}")
+
+                detector_params = {
+                    "centers": centers,
+                    "uhats": uhats,
+                    "vhats": vhats,
+                    "m": m,
+                    "n": n,
+                    "pw": pw,
+                    "ph": ph,
+                    "modes": detector_modes,
+                    "radial_bound": detector_radial_bound_frac,
+                    "global_rot_bound_deg": detector_global_rot_bound_deg,
+                    "global_rot_axis": np.array(detector_global_rot_axis),
+                    "global_trans_bound_meters": detector_global_trans_bound_meters,
+                }
+
+            xyz_out = np.zeros((len(pixel_r), 3))
+            tt_out = np.zeros(len(pixel_r))
+            az_out = np.zeros(len(pixel_r))
+
+            u_offsets = np.zeros(len(pixel_r))
+            v_offsets = np.zeros(len(pixel_r))
+            bank_indices = np.zeros(len(pixel_r), dtype=np.int32)
+
+            for phys_bank in np.unique(bank_array):
+                mask = bank_array == phys_bank
+                if not np.any(mask):
+                    continue
+
+                try:
+                    det_config = beamlines[instrument_name][str(int(phys_bank))]
+                    det = Detector(det_config)
+
+                    xyz_p = det.pixel_to_lab(pixel_r[mask], pixel_c[mask])
+                    xyz_out[mask] = xyz_p
+
+                    tt_out[mask], az_out[mask] = det.pixel_to_angles(
+                        pixel_r[mask], pixel_c[mask], ki_vec=ki_vec_val
+                    )
+
+                    if refine_detector and int(phys_bank) in bank_to_idx:
+                        bank_indices[mask] = bank_to_idx[int(phys_bank)]
+                        u_offsets[mask] = np.dot(xyz_p - det.center, det.uhat)
+                        v_offsets[mask] = np.dot(xyz_p - det.center, det.vhat)
+
+                except KeyError as e:
+                    print(
+                        f"Warning: Could not rebuild geometry for bank {phys_bank}: {e}"
+                    )
+
+            input_data["peaks/xyz"] = xyz_out
+            input_data["peaks/two_theta"] = tt_out
+            input_data["peaks/azimuthal"] = az_out
+
+            if refine_detector:
+                peak_pixel_coords = {
+                    "u_offsets": u_offsets.tolist(),
+                    "v_offsets": v_offsets.tolist(),
+                    "bank_indices": bank_indices.tolist(),
+                }
+        else:
+            raise ValueError(
+                "ERROR: Input file does not contain peaks/pixel_r and peaks/pixel_c. Cannot perform physically sound indexing."
+            )
+
+    if "peaks/image_index" in input_data:
+        input_data["peaks/run_index"] = input_data["peaks/image_index"]
+
+    # --- INJECT SECOND PHASE OF BOOTSTRAP PHYSICS ---
+    if bootstrap_filename:
+        with h5py.File(bootstrap_filename, "r") as b_f:
+            if "sample/offset" in b_f:
+                input_data["sample/offset"] = b_f["sample/offset"][()]
+            if "beam/ki_vec" in b_f:
+                ki_vec_val = b_f["beam/ki_vec"][()]
+
+    input_data["sample/a"], input_data["sample/b"], input_data["sample/c"] = a, b, c
+    (
+        input_data["sample/alpha"],
+        input_data["sample/beta"],
+        input_data["sample/gamma"],
+    ) = alpha, beta, gamma
+    input_data["sample/space_group"] = space_group
+    input_data["instrument/wavelength"] = [float(wavelength_min), float(wavelength_max)]
+    input_data["beam/ki_vec"] = ki_vec_val
+
+    opt = FindUB(data=input_data)
+    opt.wavelength = [float(wavelength_min), float(wavelength_max)]
+
+    if bootstrap_filename:
+        with h5py.File(bootstrap_filename, "r") as b_f:
+            if "optimization/goniometer_offsets" in b_f:
+                off_data = b_f["optimization/goniometer_offsets"]
+                if isinstance(off_data, h5py.Group):
+                    opt.goniometer_offsets = {
+                        k: off_data[k][()] for k in off_data.keys()
+                    }
+                else:
+                    opt.goniometer_offsets = off_data[()]
+
+    print(f"Starting evosax optimization with strategy: {strategy_name}")
+    print(f"Running {n_runs} run(s)...")
+    print(f"Settings per run: Population Size={population_size}, Generations={gens}")
+    if freeze_orientation:
+        print("ORIENTATION LOCKED: U Matrix will not be refined.")
+    if refine_lattice:
+        print(f"Refining lattice parameters with {lattice_bound_frac * 100}% bounds.")
+    if refine_sample:
+        print(f"Refining sample offset with {1000 * sample_bound_meters} mm bounds.")
+    if refine_beam:
+        print(f"Refining beam tilt with {beam_bound_deg}° bounds.")
+
+    goniometer_names = None
+
+    if original_nexus_filename and instrument_name:
+        is_merged = False
+        with h5py.File(original_nexus_filename, "r") as f_check:
+            if "images" in f_check and "goniometer/axes" in f_check:
+                is_merged = True
+                axes = f_check["goniometer/axes"][()]
+                angles = f_check["goniometer/angles"][()]
+                names = (
+                    [n.decode("utf-8") for n in f_check["goniometer/names"][()]]
+                    if "goniometer/names" in f_check
+                    else None
+                )
+
+        if not is_merged:
+            # This forces the pipeline to read the UPDATED reduction_settings.json
+            axes, angles, names = get_rotation_data_from_nexus(
+                original_nexus_filename, instrument_name
+            )
+
+        if len(axes) > 0:
+            opt.goniometer_axes = np.array(axes)
+            angles = np.array(angles)
+            if angles.ndim == 1:
+                angles = angles.reshape(-1, 1)
+
+            opt.goniometer_angles = angles
+            goniometer_names = names
+            if names is not None:
+                opt.goniometer_names = names
+
+            # 1. Overwrite the stale baked axes/angles from finder.h5
+            input_data["goniometer/axes"] = opt.goniometer_axes
+            input_data["goniometer/angles"] = opt.goniometer_angles
+            if names is not None:
+                input_data["goniometer/names"] = [
+                    n.encode("utf-8") if isinstance(n, str) else n for n in names
+                ]
+
+            # This forces JAX VectorizedObjective to build the R matrix dynamically from the new JSON axes.
+            opt.R = None
+            if "goniometer/R" in input_data:
+                del input_data["goniometer/R"]
+
+            # --- RUN/PEAK MAPPING LOGIC ---
+            if opt.run_indices is not None:
+                max_run_id = int(np.max(opt.run_indices))
+                num_peaks = len(opt.run_indices)
+                num_axes = len(opt.goniometer_axes)
+
+                # 1. Safely orient angles to (num_axes, N) without blindly transposing square matrices
+                if angles.ndim == 2:
+                    if angles.shape[0] != num_axes and angles.shape[1] == num_axes:
+                        angles = angles.T
+                elif angles.ndim == 1:
+                    angles = angles.reshape(num_axes, 1)
+
+                num_angles_provided = angles.shape[1]
+
+                # 2. Check for single-frame tiled data from run_reduce
+                # If all columns are identical, collapse it back to a single angle.
+                if num_angles_provided > 1:
+                    if np.allclose(angles, angles[:, 0:1], atol=1e-7):
+                        angles = angles[:, 0:1]
+                        num_angles_provided = 1
+
+                # 3. Safely map angles to cover the highest requested physical index
+                if num_angles_provided == 1:
+                    # Single frame: safe to broadcast to cover any max_run_id (e.g., Bank 105)
+                    opt.goniometer_angles = np.tile(angles, (1, max_run_id + 1))
+                elif num_angles_provided > max_run_id:
+                    # Multi-frame: We have enough explicit angles to cover the highest index
+                    opt.goniometer_angles = angles
+                elif num_angles_provided == num_peaks:
+                    # Angles provided explicitly per peak
+                    opt.goniometer_angles = angles
+                else:
+                    # Multi-frame mismatch: run_indices contains physical bank IDs (e.g. 105)
+                    # but angles only contains contiguous steps (e.g. 52).
+                    print(
+                        f"WARNING: Angle shape {angles.shape} does not cover max run index {max_run_id}."
+                    )
+                    print(
+                        "Padding goniometer angles to prevent out-of-bounds lookup..."
+                    )
+                    padded_angles = np.zeros((num_axes, max_run_id + 1))
+                    padded_angles[:, :num_angles_provided] = angles
+                    for i in range(num_angles_provided, max_run_id + 1):
+                        padded_angles[:, i] = angles[:, -1]
+                    opt.goniometer_angles = padded_angles
+            else:
+                num_peaks = len(opt.two_theta) if opt.two_theta is not None else 1
+                num_axes = len(opt.goniometer_axes)
+
+                if angles.ndim == 2 and angles.shape[1] == num_axes:
+                    angles = angles.T
+
+                num_angles_provided = (
+                    angles.shape[1] if angles.ndim == 2 else len(angles)
+                )
+
+                if num_angles_provided == num_peaks:
+                    opt.goniometer_angles = angles
+                elif num_angles_provided == 1:
+                    opt.goniometer_angles = np.tile(angles, (1, num_peaks))
+                else:
+                    raise ValueError(
+                        f"CRITICAL: Angle shape {angles.shape} cannot map to {num_peaks} peaks."
+                    )
+
+    # Apply the console messages appropriately
+    if refine_goniometer:
+        print(
+            f"Refining goniometer angles from fresh JSON/Nexus with {goniometer_bound_deg} deg bounds."
+        )
+    elif opt.goniometer_axes is not None:
+        print("Using fresh kinematics from JSON (no refinement).")
+    else:
+        print("WARNING: No goniometer data found.")
+
+    init_params = None
+    if bootstrap_filename:
+        init_params = opt.get_bootstrap_params(
+            refine_goniometer_axes=refine_goniometer_axes,
+            bootstrap_filename=bootstrap_filename,
+            freeze_orientation=freeze_orientation,
+        )
+
+    num, hkl, lamda, U = opt.minimize(
+        strategy_name=strategy_name,
+        population_size=population_size,
+        num_generations=gens,
+        n_runs=n_runs,
+        sigma_init=sigma_init,
+        seed=seed,
+        init_params=init_params,
+        goniometer_bound_deg=goniometer_bound_deg,
+        refine_lattice=refine_lattice,
+        lattice_bound_frac=lattice_bound_frac,
+        refine_goniometer=refine_goniometer,
+        refine_goniometer_axes=refine_goniometer_axes,
+        goniometer_names=goniometer_names,
+        refine_sample=refine_sample,
+        sample_bound_meters=sample_bound_meters,
+        refine_beam=refine_beam,
+        beam_bound_deg=beam_bound_deg,
+        batch_size=batch_size,
+        refine_detector=refine_detector,
+        detector_params=detector_params,
+        peak_pixel_coords=peak_pixel_coords,
+        detector_trans_bound_meters=detector_trans_bound_meters,
+        detector_rot_bound_deg=detector_rot_bound_deg,
+        freeze_orientation=freeze_orientation,
+        num_candidates=num_candidates,
+    )
+
+    print(f"\nOptimization complete. Best solution indexed {num} peaks.")
+    opt.reciprocal_lattice_B()
+
+    copy_keys = [
+        "sample/space_group",
+        "instrument/wavelength",
+        "peaks/intensity",
+        "peaks/sigma",
+        "peaks/radius",
+        "goniometer/R",
+        "goniometer/axes",
+        "goniometer/angles",
+        "goniometer/names",
+        "files",
+        "file_offsets",
+        "peaks/run_index",
+        "peaks/image_index",
+        "bank",
+        "sample/offset",
+        "beam/ki_vec",
+        "peaks/pixel_r",
+        "peaks/pixel_c",
+    ]
+
+    copied_data = {}
+    for key in copy_keys:
+        if key in input_data:
+            copied_data[key] = input_data[key]
+
+    print(f"Saving indexed peaks to {output_peaks_filename}...")
+    with h5py.File(output_peaks_filename, "w") as f:
+        if instrument_name:
+            f.attrs["instrument"] = instrument_name
+        elif "instrument" in input_data:
+            f.attrs["instrument"] = input_data["instrument"]
+
+        for key, value in copied_data.items():
+            f[key] = value
+
+        def safe_write(grp, name, data):
+            if name in grp:
+                del grp[name]
+            grp[name] = data
+
+        safe_write(f, "goniometer/R", opt.R)
+
+        if opt.goniometer_offsets is not None:
+            grp_name = "optimization/goniometer_offsets"
+            if grp_name in f:
+                del f[grp_name]
+
+            if isinstance(opt.goniometer_offsets, dict):
+                grp = f.create_group(grp_name)
+                for k, v in opt.goniometer_offsets.items():
+                    grp[k] = v
+            else:
+                f[grp_name] = opt.goniometer_offsets
+
+        safe_write(f, "sample/a", opt.a)
+        safe_write(f, "sample/b", opt.b)
+        safe_write(f, "sample/c", opt.c)
+        safe_write(f, "sample/alpha", opt.alpha)
+        safe_write(f, "sample/beta", opt.beta)
+        safe_write(f, "sample/gamma", opt.gamma)
+        safe_write(f, "sample/offset", opt.sample_offset)
+
+        B_mat = opt.reciprocal_lattice_B()
+        safe_write(f, "sample/B", B_mat)
+        f["sample/U"] = U
+
+        if opt.run_indices is not None:
+            safe_write(f, "peaks/run_index", opt.run_indices)
+
+        f["peaks/h"] = hkl[:, 0]
+        f["peaks/k"] = hkl[:, 1]
+        f["peaks/l"] = hkl[:, 2]
+        f["peaks/lambda"] = lamda
+
+        if opt.x is not None and opt.x.size > 0:
+            f["optimization/best_params"] = opt.x
+
+        import json
+
+        flags = {
+            "refine_lattice": refine_lattice,
+            "refine_goniometer": refine_goniometer,
+            "refine_sample": refine_sample,
+            "refine_beam": refine_beam,
+            "refine_detector": refine_detector,
+            "freeze_orientation": freeze_orientation,
+        }
+        f.create_dataset("optimization/flags", data=json.dumps(flags).encode("utf-8"))
+
+        if refine_detector and hasattr(opt, "calibrated_centers"):
+            for b_idx, b_id in enumerate(target_banks):
+                grp_name = f"detector_calibration/bank_{b_id}"
+                f.create_group(grp_name)
+                f[f"{grp_name}/center"] = opt.calibrated_centers[b_idx]
+                f[f"{grp_name}/uhat"] = opt.calibrated_uhats[b_idx]
+                f[f"{grp_name}/vhat"] = opt.calibrated_vhats[b_idx]
+    print("Done.")
