+++ src/subhkl/io/commands/integrator.py
@@ -1,0 +1,263 @@
+import h5py
+import numpy as np
+
+from subhkl.integration import Peaks
+from subhkl.instrument.detector import calibrate_from_file
+
+def run(
+    filename: str,
+    instrument: str,
+    integration_peaks_filename: str,
+    output_filename: str,
+    integration_method: str = "free_fit",
+    integration_mask_file: str | None = None,
+    integration_mask_rel_erosion_radius: float | None = 0.05,
+    region_growth_distance_threshold: float = 1.5,
+    region_growth_minimum_intensity: float = 50.0,
+    region_growth_minimum_sigma: float | None = None,
+    region_growth_maximum_pixel_radius: float = 17.0,
+    peak_center_box_size: int = 15,
+    peak_smoothing_window_size: int = 15,
+    peak_minimum_pixels: int = 10,
+    peak_minimum_signal_to_noise: float = 1.0,
+    peak_pixel_outlier_threshold: float = 2.0,
+    create_visualizations: bool = False,
+    show_progress: bool = True,
+    found_peaks_file: str | None = None,
+    max_workers: int = 16,
+):
+    calibrate_from_file(integration_peaks_filename, instrument)
+
+    peak_dict = {}
+    angles_stack = None
+    all_R = None
+    with h5py.File(integration_peaks_filename, "r") as f:
+        U = f["sample/U"][()] if "sample/U" in f else None
+        B = f["sample/B"][()] if "sample/B" in f else None
+        all_R = f["goniometer/R"][()] if "goniometer/R" in f else None
+        angles_stack = f["goniometer/angles"][()] if "goniometer/angles" in f else None
+        sample_offset = f["sample/offset"][()] if "sample/offset" in f else np.zeros(3)
+        ki_vec = (
+            f["beam/ki_vec"][()] if "beam/ki_vec" in f else np.array([0.0, 0.0, 1.0])
+        )
+
+        for key in f["banks"].keys():
+            img_idx = int(key)
+            grp = f[f"banks/{key}"]
+            peak_dict[img_idx] = [
+                grp["i"][()],
+                grp["j"][()],
+                grp["h"][()],
+                grp["k"][()],
+                grp["l"][()],
+                grp["wavelength"][()],
+            ]
+
+    peaks = Peaks(filename, instrument)
+
+    integration_params = {
+        "region_growth_distance_threshold": region_growth_distance_threshold,
+        "region_growth_minimum_intensity": region_growth_minimum_intensity,
+        "region_growth_minimum_sigma": region_growth_minimum_sigma,
+        "region_growth_maximum_pixel_radius": region_growth_maximum_pixel_radius,
+        "peak_center_box_size": peak_center_box_size,
+        "peak_smoothing_window_size": peak_smoothing_window_size,
+        "peak_minimum_pixels": peak_minimum_pixels,
+        "peak_minimum_signal_to_noise": peak_minimum_signal_to_noise,
+        "peak_pixel_outlier_threshold": peak_pixel_outlier_threshold,
+        "integration_mask_file": integration_mask_file,
+        "integration_mask_rel_erosion_radius": integration_mask_rel_erosion_radius,
+    }
+
+    if all_R is None:
+        print("Warning: Refined R stack not found in prediction file. Using nominal.")
+        all_R = peaks.goniometer.rotation
+
+    if angles_stack is None:
+        angles_stack = peaks.goniometer.angles_raw
+
+    UB = U @ B if U is not None and B is not None else None
+    RUB = None
+    if UB is not None:
+        RUB = np.matmul(all_R, UB) if all_R.ndim == 3 else all_R @ UB
+
+    result = peaks.integrate(
+        peak_dict,
+        integration_params,
+        RUB=RUB,
+        R_stack=all_R,
+        angles_stack=angles_stack,
+        sample_offset=sample_offset,
+        ki_vec=ki_vec,
+        create_visualizations=create_visualizations,
+        show_progress=show_progress,
+        integration_method=integration_method,
+        file_prefix=filename,
+        found_peaks_file=found_peaks_file,
+        max_workers=max_workers,
+    )
+
+    print(f"Saving integrated peaks to {output_filename}")
+
+    copy_keys = [
+        "sample/a",
+        "sample/b",
+        "sample/c",
+        "sample/alpha",
+        "sample/beta",
+        "sample/gamma",
+        "sample/space_group",
+        "sample/U",
+        "sample/B",
+        "sample/offset",
+        "beam/ki_vec",
+        "instrument/wavelength",
+    ]
+
+    with h5py.File(output_filename, "w") as f:
+        f["peaks/h"], f["peaks/k"], f["peaks/l"] = result.h, result.k, result.l
+        f["peaks/lambda"] = result.wavelength
+        f["peaks/intensity"], f["peaks/sigma"] = result.intensity, result.sigma
+        f["peaks/two_theta"], f["peaks/azimuthal"] = result.tt, result.az
+        f["peaks/bank"] = result.bank
+        f["peaks/run_index"] = result.run_id
+        f["peaks/xyz"] = result.xyz
+
+        if result.R and any(r is not None for r in result.R):
+            f["goniometer/R"] = np.array(result.R)
+        if result.angles and any(a is not None for a in result.angles):
+            f["goniometer/angles"] = np.array(result.angles)
+
+        with h5py.File(integration_peaks_filename, "r") as f_in:
+            for key in copy_keys:
+                if key in f_in:
+                    f_in.copy(f_in[key], f, key)
+            for k in ["goniometer/axes", "goniometer/names"]:
+                if k in f_in:
+                    f_in.copy(f_in[k], f, k)
+
+def run_rbf(
+    filename: str,
+    instrument: str,
+    integration_peaks_filename: str,
+    output_filename: str,
+    alpha: float = 1.0,
+    gamma: float = 1.0,
+    sigmas: str = "1.0,2.0,4.0",
+    nominal_sigma: float = 1.0,
+    anisotropic: bool = False,
+    fit_mosaicity: bool = False,
+    max_peaks: int = 500,
+    rel_border_width: float = 0.0,
+    show_progress: bool = True,
+    create_visualizations: bool = False,
+    chunk_size: int = 256,
+    max_workers: int | None = None,
+):
+    calibrate_from_file(integration_peaks_filename, instrument)
+
+    import h5py
+    from subhkl.search.sparse_rbf import integrate_peaks_rbf_ssn
+
+    sigma_list = [float(k.strip()) for k in sigmas.split(",")]
+    print(f"Starting Dense Sparse RBF Integration on {filename}")
+    print(
+        f"Parameters: Alpha={alpha}, Gamma={gamma}, Sigma={sigma_list}, Max Peaks Padding={max_peaks}"
+    )
+
+    peak_dict = {}
+
+    with h5py.File(integration_peaks_filename, "r") as f:
+        if "sample/U" in f:
+            f["sample/U"][()]
+        if "sample/B" in f:
+            f["sample/B"][()]
+        if "goniometer/R" in f:
+            all_R = f["goniometer/R"][()]
+        if "goniometer/angles" in f:
+            angles_stack = f["goniometer/angles"][()]
+
+        if "sample/offset" in f:
+            sample_offset = f["sample/offset"][()]
+        else:
+            sample_offset = np.zeros(3)
+
+        for key in f["banks"].keys():
+            img_idx = int(key)
+            grp = f[f"banks/{key}"]
+            peak_dict[img_idx] = [
+                grp["i"][()],
+                grp["j"][()],
+                grp["h"][()],
+                grp["k"][()],
+                grp["l"][()],
+                grp["wavelength"][()],
+            ]
+
+    peaks = Peaks(filename, instrument)
+
+    if all_R is None:
+        all_R = peaks.goniometer.rotation
+    if angles_stack is None:
+        angles_stack = peaks.goniometer.angles_raw
+
+    one_image = next(iter(peaks.image.ims.values()))
+    border_width = int(rel_border_width * min(one_image.shape[0], one_image.shape[1]))
+
+    result = integrate_peaks_rbf_ssn(
+        peak_dict=peak_dict,
+        peaks_obj=peaks,  # Pass the full Peaks object
+        alpha=alpha,
+        sigmas=sigma_list,
+        gamma=gamma,
+        nominal_sigma=nominal_sigma,
+        max_peaks=max_peaks,
+        show_progress=show_progress,
+        all_R=all_R,  # Pass rotation and offset downstream
+        sample_offset=sample_offset,
+        anisotropic=anisotropic,
+        fit_mosaicity=fit_mosaicity,
+        border_width=border_width,
+        chunk_size=chunk_size,
+        create_visualizations=create_visualizations,
+        file_prefix=filename,
+        max_workers=max_workers,
+    )
+
+    print(f"Saving RBF integrated peaks to {output_filename}")
+    with h5py.File(output_filename, "w") as f:
+        f["peaks/h"] = result.h
+        f["peaks/k"] = result.k
+        f["peaks/l"] = result.l
+        f["peaks/lambda"] = result.wavelength
+        f["peaks/intensity"] = result.intensity
+        f["peaks/sigma"] = result.sigma  # SVD-stabilized Fisher Info UQ
+        f["peaks/two_theta"] = result.tt
+        f["peaks/azimuthal"] = result.az
+        f["peaks/bank"] = result.bank
+        f["peaks/run_index"] = result.run_id
+
+        # Copy full metadata context from predictor output
+        copy_keys = [
+            "sample/a",
+            "sample/b",
+            "sample/c",
+            "sample/alpha",
+            "sample/beta",
+            "sample/gamma",
+            "sample/space_group",
+            "sample/U",
+            "sample/B",
+            "sample/offset",
+            "beam/ki_vec",
+            "instrument/wavelength",
+        ]
+
+        with h5py.File(integration_peaks_filename, "r") as f_in:
+            for key in copy_keys:
+                if key in f_in:
+                    f_in.copy(f_in[key], f, key)
+
+            for k in ["goniometer/axes", "goniometer/names"]:
+                if k in f_in:
+                    f_in.copy(f_in[k], f, k)
