+++ src/subhkl/io/commands/merge_images.py
@@ -1,0 +1,49 @@
+import h5py
+import glob
+
+from subhkl.core.spacegroup import get_space_group_object
+from subhkl.io.export import ImageStackMerger 
+
+def run(
+    input_pattern: str,
+    output_filename: str,
+    a: float,
+    b: float,
+    c: float,
+    alpha: float,
+    beta: float,
+    gamma: float,
+    space_group: str,
+):
+
+    try:
+        get_space_group_object(space_group)
+    except ValueError as e:
+        raise ValueError(f"ERROR: Invalid space group '{space_group}': {e}")
+
+    if " " in input_pattern:
+        h5_files = []
+        for p in input_pattern.split():
+            h5_files.extend(glob.glob(p))
+    else:
+        h5_files = glob.glob(input_pattern)
+
+    h5_files = sorted(list(set(h5_files)))
+
+    if not h5_files:
+        raise ValueError(f"No files found matching: {input_pattern}")
+
+    print(f"Found {len(h5_files)} files. Merging...")
+    merger = ImageStackMerger(h5_files)
+    merger.merge(output_filename)
+
+    with h5py.File(output_filename, "a") as f:
+        f["sample/a"] = a
+        f["sample/b"] = b
+        f["sample/c"] = c
+        f["sample/alpha"] = alpha
+        f["sample/beta"] = beta
+        f["sample/gamma"] = gamma
+        f["sample/space_group"] = space_group.encode("utf-8")
+
+    print(f"Successfully created {output_filename} with unit cell info embedded.")
