+++ src/subhkl/io/commands/metrics.py
@@ -1,0 +1,45 @@
+import numpy as np
+from typing import List
+from subhkl.instrument.metrics import compute_metrics
+
+def run(
+    file1: str,
+    file2: str | None = None,
+    instrument: str | None = None,
+    d_min: float | None = None,
+    per_run: bool = False,
+    ki_vec: List[float] | np.ndarray = None,
+):
+
+    # No need to call calibrate_from_file here because metrics.py
+    # dynamically shifts coordinates using the detector_calibration group.
+    result = compute_metrics(
+        file1=file1,
+        file2=file2,
+        instrument=instrument,
+        d_min=d_min,
+        per_run=per_run,
+        ki_vec_override=ki_vec,
+    )
+
+    if "error_message" in result:
+        print(result["error_message"])
+        if result["error_message"].startswith("Exception"):
+            print("METRICS: 9.99 9.99 9.99 9.99 9.99 9.99")
+        return
+
+    if "filter_message" in result:
+        print(f"METRICS: {result['filter_message']}")
+
+    # Print main metrics
+    print(
+        f"METRICS: {result['median_d_err']:.5f} {result['mean_d_err']:.5f} {result['max_d_err']:.5f} "
+        f"{result['median_ang_err']:.5f} {result['mean_ang_err']:.5f} {result['max_ang_err']:.5f}"
+    )
+
+    # Print per-run metrics if requested
+    if per_run and "per_run_errors" in result:
+        print("\nPER-RUN MEDIAN ANGULAR ERROR (deg) - Sorted by error:")
+        for r, err, count in result["per_run_errors"]:
+            status = "BAD" if err > 1.0 else "OK"
+            print(f"  Run {r:4d}: {err:6.3f} ({count:4d} peaks) [{status}]")
