+++ src/subhkl/io/commands/reduce.py
@@ -1,0 +1,59 @@
+import h5py
+import numpy as np
+
+from subhkl.integration import Peaks
+
+def run(
+    nexus_filename: str,
+    output_filename: str,
+    instrument: str,
+    wavelength_min: float | None = None,
+    wavelength_max: float | None = None,
+):
+    print(f"Reducing {nexus_filename} -> {output_filename}")
+    peaks_handler = Peaks(
+        nexus_filename,
+        instrument,
+        wavelength_min=wavelength_min,
+        wavelength_max=wavelength_max,
+    )
+
+    if not peaks_handler.image.ims:
+        print("Warning: No images found in file.")
+        return
+
+    sorted_banks = sorted(peaks_handler.image.ims.keys())
+    image_stack = np.stack([peaks_handler.image.ims[b] for b in sorted_banks])
+    bank_ids = np.array(sorted_banks, dtype=np.int32)
+    n_images = len(sorted_banks)
+
+    if peaks_handler.goniometer.angles_raw is not None:
+        angles_repeated = np.tile(peaks_handler.goniometer.angles_raw, (n_images, 1))
+    else:
+        angles_repeated = np.zeros((n_images, 3))
+
+    axes = (
+        np.array(peaks_handler.goniometer.axes_raw)
+        if peaks_handler.goniometer.axes_raw is not None
+        else np.array([0.0, 1.0, 0.0])
+    )
+
+    with h5py.File(output_filename, "w") as f:
+        f.create_dataset("images", data=image_stack, compression="lzf")
+        f.create_dataset("bank_ids", data=bank_ids)
+        f.create_dataset("goniometer/angles", data=angles_repeated)
+        f.create_dataset("goniometer/axes", data=axes)
+
+        if peaks_handler.goniometer.names_raw:
+            dt = h5py.string_dtype(encoding="utf-8")
+            f.create_dataset(
+                "goniometer/names", data=peaks_handler.goniometer.names_raw, dtype=dt
+            )
+
+        f.create_dataset(
+            "instrument/wavelength",
+            data=[peaks_handler.wavelength.min, peaks_handler.wavelength.max],
+        )
+        f.attrs["instrument"] = instrument
+
+    print(f"Saved {n_images} banks to {output_filename}")
